move authentication classes to authentication module
