+++ authentication/src/main/AndroidManifest.xml
@@ -1,0 +1,14 @@
+<manifest package="com.hodamohammadi.authentication"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <application
+        android:allowBackup="true"
+        android:fullBackupContent="true">
+        <activity android:name="com.hodamohammadi.authentication.AuthenticationActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <action android:name="android.intent.action.VIEW" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+    </application>
+</manifest>
