+++ authentication/src/main/kotlin/com/hodamohammadi/authentication/AuthenticationActivity.kt
@@ -1,10 +1,10 @@
-package com.hodamohammadi.conveyor.activities
+package com.hodamohammadi.authentication
 
 import android.support.v7.app.AppCompatActivity
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
-import com.hodamohammadi.conveyor.R
+import com.hodamohammadi.authentication.R
 import com.firebase.ui.auth.AuthUI
 import com.firebase.ui.auth.IdpResponse
 import com.hodamohammadi.commons.navigation.loaders.loadIntentOrNull
