+++ authentication/src/main/res/layout/authentication_activity.xml
