+++ navigation/src/main/AndroidManifest.xml
@@ -1,2 +1,1 @@
-<manifest
-    package="com.hodamohammadi.navigation" />
+<manifest package="com.hodamohammadi.navigation" />
